rent_home/requirements.txt
Django==1.10.4
djangorestframework==3.5.3
mysqlclient==1.3.12
